refactor(cronjobs): update HttpClient import path and remove obsolete files
- Update import in tor_exit_nodes.py and whatsmyip.py from
  greedybear.cronjobs.client to greedybear.cronjobs.http_client
- Fix whatsmyip error handling to use self.log.exception() for consistency
- Remove obsolete greedybear/cronjobs/client.py (superseded by http_client.py)
- Remove obsolete tests/test_client.py (superseded by test_http_client.py)

Refs #1214
